Added Java practice programs

+++ Java_Practice_Programs/practicePrograms/src/Day2/Factorials.java
@@ -10,7 +10,7 @@ public class Factorials {
 		// n! = n*(n-1)*(n-2)*(n-3).....
 		Scanner sc  = new Scanner(System.in);
 		System.out.println("Enter number Factorial for ");
-		int a=1; 
+		double a=1; 
 		int num = sc.nextInt();
 		
 		for(int i=2;i<=num;i++)
